optimised centroid tests. added subtract operator overload for pixel_coordinates struct. astroUtilities to make a uniform vector, and fixed bug when calculating stDev. astroUtilities method to check size of vectors added. added MonteCarlo class, with multithread architecture. added some methods on FrameProcessor to find centroid with recursive methods. FIxed bug and improved methods for saturated sources in Frame Class.

+++ src/Frame.hpp
@@ -138,7 +138,7 @@ private:
   ulong_t pixel_biasCounts() { return biasCounts(bias_generator); }
 
   std::vector<source> sources;
-  uint16_t h, w, hsim, wsim;
+  uint32_t h, w, hsim, wsim;
   Grid<uint32_t> simfr, fr;
 
   uint16_t nsources() { return sources.size(); }
